requirements.txt
langchain==1.2.10
langchain-text-splitters==1.0.0
langchain-ollama==1.0.1
langchain-chroma==1.1.0
pypdf==6.1.3
python-docx==1.2.0
